+++ built/grammar/GrammarConstants.d.ts
@@ -19,6 +19,7 @@ declare enum GrammarConstants {
     parseWith = "parseWith",
     keywords = "keywords",
     columns = "columns",
+    catchAllColumn = "catchAllColumn",
     catchAllKeyword = "catchAllKeyword",
     defaults = "defaults",
     constants = "constants",